Fixed a bug on dialing from recent history
Fixed a bug where if user clicks on the recently dialed
history from the carousel, and the entry was not in the
contacts list, then it will not dial since it couldn't
find the right contact.  This patch will create a fake
contact object for any known entries.  Also fixed a typo.

Change-Id: Ia623264e74c0337ff4c5be9bad84eb3887232bcb
Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>