Fix TC-2267 - Calling history refresh too slow
Phone call history is not updated efficiently since everytime a call
entry is added, it removes all the existing entries and reload
all call history items and recreates all the html elements.  It's a
huge performance hit on the app performance as the history grow longer.
This patch will only append the new call history to the existing
call history list

Change-Id: I657379b5ef1f84221b478dcc9f6530ba026d9c26
Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>